requirements.txt
appdirs==1.4.4
attrs==20.3.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.5
click==7.1.2
cycler==0.10.0
djitellopy2==2.3
fastseg==0.1.2
geffnet==1.0.0
kiwisolver==1.3.1
matplotlib==3.4.2
mypy-extensions==0.4.3
numpy==1.19.2
olefile==0.46
opencv-python==4.5.1.48
pathspec==0.7.0
Pillow==8.2.0
pip==21.0.1
pycparser==2.20
pygame==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.4.1
regex==2021.4.4
setuptools==52.0.0.post20210125
six==1.15.0
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wheel==0.36.2
yacs==0.1.8
timm==0.3.2
ipython==7.25.0
opencv-python